ruber_project/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
qrcode>=7.4.2
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8